Consolidate store declarations (#1458)
