+++ .changeset/poor-fishes-compare.md
@@ -1,0 +1,5 @@
+---
+'houdini-svelte': patch
+---
+
+Avoid errors with module-scope class dependencies
