+++ e2e/svelte/.graphqlrc.yaml
@@ -1,0 +1,9 @@
+projects:
+    default:
+        schema:
+            - ./schema.graphql
+            - ./$houdini/graphql/schema.graphql
+        documents:
+            - '**/*.gql'
+            - '**/*.svelte'
+            - ./$houdini/graphql/documents.gql
