+++ e2e/svelte/README.md
@@ -1,0 +1,1 @@
+# Houdini Svelte End-to-End
