+++ e2e/svelte/index.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+	<head>
+		<meta charset="UTF-8" />
+		<link rel="icon" type="image/png" href="https://houdinigraphql.com/images/logo.png" />
+		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
+		<title>Houdini • e2e • Svelte</title>
+	</head>
+	<body>
+		<div id="app"></div>
+		<script type="module" src="/src/main.ts"></script>
+	</body>
+</html>
