+++ e2e/svelte/src/client.ts
@@ -1,0 +1,15 @@
+import { HoudiniClient } from '$houdini'
+
+export default new HoudiniClient({
+	url: 'http://localhost:4000/graphql',
+
+	// uncomment this to configure the network call (for things like authentication)
+	// for more information, please visit here: https://www.houdinigraphql.com/guides/authentication
+	// fetchParams({ session }) {
+	//     return {
+	//         headers: {
+	//             Authentication: `Bearer ${session.token}`,
+	//         }
+	//     }
+	// }
+})
