+++ e2e/svelte/src/main.ts
@@ -1,0 +1,8 @@
+import App from './App.svelte'
+import './app.css'
+
+const app = new App({
+	target: document.getElementById('app'),
+})
+
+export default app
