+++ packages/houdini-svelte/src/plugin/config.ts
@@ -75,10 +75,10 @@ export function plugin_config(config: Config): Required<HoudiniSvelteConfig> {
 			mutation: '../runtime/stores/mutation.MutationStore',
 			fragment: '../runtime/stores/fragment.FragmentStore',
 			subscription: '../runtime/stores/subscription.SubscriptionStore',
-			queryCursor: '../runtime/stores/pagination/query.QueryStoreCursor',
-			queryOffset: '../runtime/stores/pagination/query.QueryStoreOffset',
-			fragmentCursor: '../runtime/stores/pagination/fragment.FragmentStoreCursor',
-			fragmentOffset: '../runtime/stores/pagination/fragment.FragmentStoreOffset',
+			queryCursor: '../runtime/stores/query.QueryStoreCursor',
+			queryOffset: '../runtime/stores/query.QueryStoreOffset',
+			fragmentCursor: '../runtime/stores/fragment.FragmentStoreCursor',
+			fragmentOffset: '../runtime/stores/fragment.FragmentStoreOffset',
 			...cfg?.customStores,
 		},
 	}
