+++ packages/houdini-svelte/src/runtime/fragments.ts
@@ -2,7 +2,7 @@ import type { Fragment, FragmentArtifact } from '$houdini/runtime/lib/types'
 import type { Readable } from 'svelte/store'
 
 import type { FragmentStore } from './stores'
-import type { FragmentStorePaginated } from './stores/pagination/fragment'
+import type { FragmentStorePaginated } from './stores/fragment'
 
 // function overloads meant to only return a nullable value
 // if the reference type was nullable
