--- packages/houdini-svelte/src/runtime/stores/pagination/index.ts
@@ -1,2 +1,0 @@
-export { FragmentStoreCursor, FragmentStoreOffset } from './fragment'
-export { QueryStoreCursor, QueryStoreOffset } from './query'
